Move Notification entrypoints code to react root
