--- react/NotificationBar.js
@@ -1,3 +1,0 @@
-import NotificationBar from './components/Notification/NotificationBar'
-
-export default NotificationBar
