--- react/NotificationInline.js
@@ -1,3 +1,0 @@
-import NotificationInline from './components/Notification/NotificationInline'
-
-export default NotificationInline
